Add files via upload
Created replicated health bars for each player and a blueprint for how the cool downs could theoretically work (still need the first weapon before we can know for sure). A widget for the health and cool down both is also made. The cool down bar has no functionality as of right now (waiting on a functional weapon).